Create members file and add about page data
