+++ src/data/events.js
@@ -39,97 +39,3 @@ export const pastEvents = [
     image: '/assets/event-image.png',
   },
 ];
-
-export const values = [
-  {
-      icon:"/icon6.jpeg",
-      name:"EQUITY",
-      text:"Lorem ipsum lorem ipsum lorem ipsum lorem ipsum",
-  },
-  {
-      icon:"/icon4.jpeg",
-      name:"ACCOUNTABILITY",
-      text:"Lorem ipsum lorem ipsum lorem ipsum lorem ipsum",
-  },
-  {
-      icon:"icon1.jpeg",
-      name:"COMMUNITY",
-      text:"Lorem ipsum lorem ipsum lorem ipsum lorem ipsum",
-  },
-  {
-      icon:"icon3.jpeg",
-      name:"TEAMWORK",
-      text:"Lorem ipsum lorem ipsum lorem ipsum lorem ipsum",
-  },
-  {
-      icon:"icon2.jpeg",
-      name:"KINDNESS",
-      text:"Lorem ipsum lorem ipsum lorem ipsum lorem ipsum",
-  }
-]
-
-
-export const boardMembers = [
-  {
-    id: 1,
-    name: 'JAMES W. M',
-    title: 'MANAGING DIRECTOR',
-    text: 'Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem ',
-    image: '/profile2.jpeg',
-  },
-  {
-    id: 2,
-    name: 'MARY SUSAN',
-    title: 'CREATIVE DIRECTOR',
-    text: 'Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem ',
-    image: '/profile1.jpeg',
-  },
-  {
-    id: 3,
-    name: 'NANCY W. M',
-    title: 'MANAGING DIRECTOR',
-    text: 'Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem ',
-    image: '/profile1.jpeg',
-  },
-  {
-    id: 4,
-    name: 'EDWARD K JR',
-    title: 'MANAGING DIRECTOR',
-    text: 'Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem Ipsum Lorem IPsum Lorem Ipsum Lorem ',
-    image: '/profile2.jpeg',
-  },
-]
-
-
-export const allMembers = [
-  {
-    id: 1,
-    name: 'JAMES W. M',
-    text: 'Lorem IPsum Lorem',
-    image: '/profile1.jpeg',
-  },
-  {
-    id: 2,
-    name: 'JAMES W. M',
-    text: 'Lorem IPsum Lorem',
-    image: '/profile1.jpeg',
-  },
-  {
-    id: 3,
-    name: 'JAMES W. M',
-    text: 'Lorem IPsum Lorem',
-    image: '/profile2.jpeg',
-  },
-  {
-    id: 4,
-    name: 'JAMES W. M',
-    text: 'Lorem IPsum Lorem',
-    image: '/profile1.jpeg',
-  },
-  {
-    id: 5,
-    name: 'JAMES W. M',
-    text: 'Lorem IPsum Lorem',
-    image: '/profile2.jpeg',
-  },
-]
